Update GitHub Action to stage site for all PRs to cfug/flutter.cn
This commit modifies the 'Stage site on Firebase' step in the GitHub Action workflow. Now, the step will run for all pull requests targeting the cfug/flutter.cn repository, regardless of the PR's origin. This change ensures that every change proposed to the repository goes through the same staging and link checking process.